requirements.txt
numpy
tensorflow>=2.4.1
tensorflow-gpu
tensorflow-datasets
gin-config
tqdm
scipy
librosa
h5py